requirements.txt
social-auth-app-django==1.2.0